src/requirements.txt
#cloudshell-traffic-teravm-chassis>=1.0,<1.1
cloudshell-core>=2.2,<2.3
cloudshell-shell-core>=3.1,<3.2
cloudshell-networking-devices>=2.2,<2.3
requests==2.18.4
